requirements.txt
fastapi # api framework
uvicorn # web server runtime
pydantic # for data validation
google-generativeai
python-jose[cryptography]